Adds mock repositories and tests.
Bit of a big one here guys - sorry to inflict that on you. This
adds mock repositories for storing player & game data. These
will be replaced by implementations that hit DynamoDB shortly.

There's a bit of a quirk in this intermediary commit, that fetching
games and players a second time returns an empty array. This is
because it's looking at the latest fetched players and seeing that
there's nothing new to fetch. Following commit will demonstrate
the change here better.