Latest game saved is only updated AFTER saving.
Previously if we'd ran into an issue saving the game, we'd already
have updated our latest persisted game, which technically isn't true.
Now we save the game then update this.

I haven't written a unit test to cover this case as this behaviour is
soon to be disappearing. In the near future (when the DB is hooked up)
we won't actually be saving the latest player, we'll be querying that info
from the DB instead.