Adds league table endpoint
Note that in the league table I'm 'unsafely' accessing
properties on an object. However, since the GameColor enum
fromString method returns an optional we're checking and logging
as we might as well.

I'm trying not to be too defensive while programming here, we should
assume that once we're handling game data that's been downloaded,
it should be safe, so we shouldn't null check everything.

I'll add to our item validator in the following commit so we can be sure
that we are dealing with safe objects internally.

Opinions on this wanted!