Adds thin player model wrapping JSON data
We were accessing our Lichess json data in multiple places in our
application. This is a source of change and is in multiple places
throughout the application. We still want to provide the full data to
our consumers so we wrap this data and provide accessors into it
as part of our model.

Result: changes by Lichess.com only affect the lichess class (and
our consumers - but who cares about them anyway?)